pyproject.toml
[tool.poetry]
name = "acfc"
version = "2.3.0"
description = "Alacritty Colors to Foot Colors (Converter)"
authors = ["grtcdr <ba.tahaaziz@gmail.com>"]
readme= "README.md"
classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
   "Topic :: Terminals :: Terminal Emulators/X Terminals",
   "Topic :: Utilities",
   "Development Status :: 5 - Production/Stable",
]

[tool.poetry.dependencies]
python = "^3.9"
configparser = "^5.0.2"
pyaml = "^21.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
